refactor(server): extract topology_queries.go and topology_autodetect.go
Move topology building, auto-detection, replication inference, and all
related helpers from datastore.go into two new files:

- topology_queries.go: Public API (GetClusterTopology, RefreshClusterAssignments),
  types (TopologyServerInfo, TopologyGroup, TopologyCluster, etc.), hierarchy
  builders, and visibility pruning functions.

- topology_autodetect.go: Auto-detection logic (buildAutoDetectedClusters),
  replication extractors (extractBinaryRelationships, extractSpockRelationships,
  extractLogicalRelationships), Spock/logical grouping, persisted member handling,
  and cluster assignment synchronization.

This is the largest single extraction in the decomposition effort, moving 50+
tightly coupled topology functions spanning ~2,300 lines. The split keeps
the public API and hierarchy builders in topology_queries.go while moving
the more internal auto-detection machinery to topology_autodetect.go.

Part of #75.
